+++ source/extensions/omni.isaac.diff_usd/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.0] - 2022-09-13
+
+### Changed
+- Moved to isaac utils menu so that it can be cleaned up on shutdown properly
+
+
 ## [0.1.0] - 2022-03-30
 
 ### Added